+++ client/src/pages/Home.jsx
@@ -1,4 +1,5 @@
-import TopSongGallary from "../components/TopSongGallary";
+import TopSongGallary from "../components/TopSongGallary/TopSongGallary";
+
 
 export default function Home() {
   return <div> <TopSongGallary/></div>;
