+++ client/src/pages/SongPage.jsx
@@ -1,8 +1,8 @@
-import TopSongGallary from "../components/TopSongGallary"
+
 function SongPage() {
   return (
     <>
-    <TopSongGallary/>
+
     </>
   );
 }
